requirements.txt
flask==2.0.1
werkzeug==2.0.3
tushare==1.2.89
numpy>=1.22.4
pandas>=2.2.3
python-dotenv==0.19.0
requests==2.26.0
yfinance>=0.2.36 